Refactor CloseableProcessorSupplier
